requirements.txt
bottle==0.12.16
certifi==2018.11.29
chardet==3.0.4
cognitive-face==1.4.2
dlib==19.16.0
idna==2.8
numpy==1.15.4
opencv-python==4.0.0.21
Pillow==5.3.0
pkg-resources==0.0.0
requests==2.21.0
urllib3==1.24.1
